backend/requirements.txt
# ============================================================
# 업비트 자동매매 시스템 v3.0 — Python 의존성 목록
# ============================================================

# ─── 웹 프레임워크 ───
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ─── 데이터베이스 (PostgreSQL + 비동기 ORM) ───
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ─── 인증 / 보안 ───
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
cryptography==42.0.4

# ─── 데이터 검증 ───
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# ─── 비동기 작업 큐 (Celery + Redis) ───
celery==5.3.6
redis==5.0.1

# ─── HTTP 클라이언트 (거래소 API 호출) ───
httpx[http2]==0.27.0
aiohttp==3.9.3
websockets==12.0

# ─── 트레이딩 분석 라이브러리 ───
pandas==2.2.0
numpy==1.26.4
ta==0.11.0

# ─── 업비트 / 바이낸스 API 래퍼 ───
pyupbit==0.2.33
python-binance==1.0.19

# ─── 텔레그램 알림 ───
python-telegram-bot==21.2

# ─── 로깅 / 모니터링 ───
loguru==0.7.2

# ─── CORS / 미들웨어 ───
slowapi==0.1.9

# ─── 테스트 (로컬 개발용, Docker 빌드에는 불필요) ───
# pytest==8.0.1
# pytest-asyncio==0.23.4
